api/src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==1.11.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
elastic-transport==8.11.0
elasticsearch[async]==8.11.1
fastapi==0.109.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
loguru==0.7.2
multidict==6.0.4
opentelemetry-api==1.23.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-sdk==1.23.0
pydantic-core==2.14.6
pydantic-settings==2.1.0
pydantic==2.5.3
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.1
sniffio==1.3.0
starlette==0.35.1
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
yarl==1.9.4
